feat: lock during worker pool cleanup
- Only run one cleanup per pool at a time
- Move pool cleaner logic to its own file

Resolve BE-2343
